+++ features/step_definitions/config_steps.rb
@@ -3,7 +3,7 @@ Given(/^that metric has a default type of "(.*?)"$/) do |view|
   @hash[@metric.name] ||= {}
   @hash[@metric.name]['defaults'] ||= {}
   @hash[@metric.name]['defaults']['type'] = view
-  allow_any_instance_of(Bothan).to receive(:metrics_config) {
+  allow_any_instance_of(Bothan::App).to receive(:metrics_config) {
     @hash
   }
 end
@@ -13,7 +13,7 @@ Given(/^that metric has a default datetime of "(.*?)"$/) do |datetime|
   @hash[@metric.name] ||= {}
   @hash[@metric.name]['defaults'] ||= {}
   @hash[@metric.name]['defaults']['datetime'] = datetime
-  allow_any_instance_of(Bothan).to receive(:metrics_config) {
+  allow_any_instance_of(Bothan::App).to receive(:metrics_config) {
     @hash
   }
 end
