+++ spec/app_spec.rb
@@ -1,4 +1,4 @@
-describe Bothan do
+describe Bothan::App do
 
   context 'sets the correct visualisation types' do
