+++ spec/dcat/app_spec.rb
@@ -1,4 +1,4 @@
-describe Bothan do
+describe Bothan::App do
   before :all do
     create_data
